Implemented MyBatis, removed all JDBC implementation, renamed & moved around files
