--- src/main/java/org/example/dao/RouteDAOImpl.java
@@ -1,120 +1,0 @@
-package org.example.dao;
-
-import org.example.configurations.ConnectionPool;
-import org.example.model.Point;
-import org.example.model.Route;
-
-import java.sql.*;
-import java.util.ArrayList;
-import java.util.List;
-
-
-public class RouteDAOImpl implements RouteDAO {
-    private static final ConnectionPool CONNECTIONPOOL = ConnectionPool.getInstance();
-
-    public RouteDAOImpl() {
-    }
-    @Override
-    public void insertRoute(Route route) {
-        Connection connection = CONNECTIONPOOL.getConnectionFromPool();
-        try (
-             PreparedStatement statement = connection.prepareStatement("INSERT INTO routes (start_point, end_point, distance) VALUES (?, ?, ?)", Statement.RETURN_GENERATED_KEYS)) {
-            statement.setInt(1, route.getStartPoint().getId());
-            statement.setInt(2, route.getEndPoint().getId());
-            statement.setLong(3, route.getDistance());
-            statement.executeUpdate();
-
-            try (ResultSet generatedKeys = statement.getGeneratedKeys()) {
-                while (generatedKeys.next()) {
-                    int id = generatedKeys.getInt(1);
-                    route.setId(id);
-                    System.out.println("Route inserted successfully. Generated ID: " + id);
-                }
-            }
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            CONNECTIONPOOL.releaseConnectionToPool(connection);
-        }
-    }
-
-    @Override
-    public void updateRoute(Route route) {
-        Connection connection = CONNECTIONPOOL.getConnectionFromPool();
-        try (
-             PreparedStatement statement = connection.prepareStatement("UPDATE routes SET distance = ? WHERE start_point = ? AND end_point = ?")) {
-            statement.setLong(1, route.getDistance());
-            statement.setInt(2, route.getStartPoint().getId());
-            statement.setInt(3, route.getEndPoint().getId());
-            int rowsAffected = statement.executeUpdate();
-            if (rowsAffected > 0) {
-                System.out.println("Route updated successfully.");
-            } else {
-                System.out.println("Route not found.");
-            }
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            CONNECTIONPOOL.releaseConnectionToPool(connection);
-        }
-    }
-
-    @Override
-    public Route getRoute(int id) {
-        Connection connection = CONNECTIONPOOL.getConnectionFromPool();
-        try (
-             PreparedStatement statement = connection.prepareStatement("SELECT r.start_point AS startPointId, r.end_point AS endPointId, r.distance AS shortestDistance FROM routes r WHERE id = ?")) {
-            statement.setInt(1, id);
-            try (ResultSet resultSet = statement.executeQuery()) {
-                if (resultSet.next()) {
-                    int startPointId = resultSet.getInt("start_point");
-                    int endPointId = resultSet.getInt("end_point");
-                    long shortestDistance = resultSet.getLong("distance");
-
-                    PointDAO pointDAO = new PointDAOImpl();
-                    Point startPoint = pointDAO.getPoint(startPointId);
-                    Point endPoint = pointDAO.getPoint(endPointId);
-
-                    Route route = new Route(startPoint, endPoint, shortestDistance);
-                    route.setId(id);
-                    return route;
-                }
-            }
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            CONNECTIONPOOL.releaseConnectionToPool(connection);
-        }
-        return null;
-    }
-
-    @Override
-    public List<Route> getRoutes() {
-        List<Route> routes = new ArrayList<>();
-        Connection connection = CONNECTIONPOOL.getConnectionFromPool();
-        try (
-             PreparedStatement statement = connection.prepareStatement("SELECT r.id AS routesId, r.start_point AS startPointId, r.end_point AS endPointId, r.distance AS shortestDistance FROM routes r");
-             ResultSet resultSet = statement.executeQuery()) {
-            PointDAO pointDAO = new PointDAOImpl();
-            while (resultSet.next()) {
-                int id = resultSet.getInt("id");
-                int startPointId = resultSet.getInt("start_point");
-                int endPointId = resultSet.getInt("end_point");
-                long shortestDistance = resultSet.getLong("distance");
-
-                Point startPoint = pointDAO.getPoint(startPointId);
-                Point endPoint = pointDAO.getPoint(endPointId);
-
-                Route route = new Route(startPoint, endPoint, shortestDistance);
-                route.setId(id);
-                routes.add(route);
-            }
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }finally {
-            CONNECTIONPOOL.releaseConnectionToPool(connection);
-        }
-        return routes;
-    }
-
-}
